+++ CHANGELOG.md
@@ -14,7 +14,8 @@ Possible header types:
 
 ## v0.6.0 (2023-11-09)
 - Improved `rotate` for tile_map and added `with_rotate` and `with_flip`
-- Added `is_subset`, `is_superset` and `symmetric difference` for tile_set
+- Added `is_subset`, `is_superset` and `symmetric difference` for tile_set and `tile_set256`
+- Added `ALL` and `is_empty` and `with_bit_set` to `tile_set256`
 
 ## v0.5.0 (2023-11-06)
 - Added `is_edge` and `is_corner` and `adjacent_tile_count` to tile